Reorganize tests for readability.
